requirements.txt
itk
monai==0.4.0
numpy
pandas
pytest
pytest-flake8
pytest-pythonpath
pytest-check
pytest-cov
scikit-learn
sklearn
tqdm
pytorch-ignite
matplotlib
pydicom
medpy
opencv-python
psutil
pydicom
medpy